update: move linking step after optimistic lock to avoid useless linkage

+++ genshin-map-api/genshin-map-api-core/genshin-map-api-core-core/src/main/java/site/yuanshen/genshin/core/service/impl/MarkerServiceImpl.java
@@ -236,15 +236,18 @@ public class MarkerServiceImpl implements MarkerService {
             }
     )
     public Boolean updateMarker(MarkerSingleDto markerSingleDto) {
-        if (markerSingleDto.getItemList() != null && !markerSingleDto.getItemList().isEmpty()) {
-            markerItemLinkMapper.delete(Wrappers.<MarkerItemLink>lambdaQuery().eq(MarkerItemLink::getMarkerId, markerSingleDto.getId()));
-            List<MarkerItemLink> itemLinkList = markerSingleDto.getItemList().parallelStream().map(markerItemLinkDto -> markerItemLinkDto.getEntity().setMarkerId(markerSingleDto.getId())).collect(Collectors.toList());
-            markerItemLinkMBPService.saveBatch(itemLinkList);
-        } else if(markerSingleDto.getItemList() != null) {
-            markerItemLinkMapper.delete(Wrappers.<MarkerItemLink>lambdaQuery().eq(MarkerItemLink::getMarkerId, markerSingleDto.getId()));
-        }
-        return markerMapper.update(markerSingleDto.getEntity(), Wrappers.<Marker>lambdaUpdate()
+        Boolean updated = markerMapper.update(markerSingleDto.getEntity(), Wrappers.<Marker>lambdaUpdate()
                 .eq(Marker::getId, markerSingleDto.getId())) == 1;
+        if(updated) {
+            if (markerSingleDto.getItemList() != null && !markerSingleDto.getItemList().isEmpty()) {
+                markerItemLinkMapper.delete(Wrappers.<MarkerItemLink>lambdaQuery().eq(MarkerItemLink::getMarkerId, markerSingleDto.getId()));
+                List<MarkerItemLink> itemLinkList = markerSingleDto.getItemList().parallelStream().map(markerItemLinkDto -> markerItemLinkDto.getEntity().setMarkerId(markerSingleDto.getId())).collect(Collectors.toList());
+                markerItemLinkMBPService.saveBatch(itemLinkList);
+            } else if (markerSingleDto.getItemList() != null) {
+                markerItemLinkMapper.delete(Wrappers.<MarkerItemLink>lambdaQuery().eq(MarkerItemLink::getMarkerId, markerSingleDto.getId()));
+            }
+        }
+        return updated;
     }
 
     /**
